Stop cleaning dist_chrome on every build
Commit 497c568cef77db519a3b1d423a44003a23ac267b started to clean
dist_chrome before every build so we can detect missing vendor files
instead of them staying permanently on a developer's machine, while
failing on new clones.

The problem is that it is not possible to block the extension refresh
during build, which is causing very annoying errors.

I kept cleaning firefox_dist so we can at least detect missing files
when testing Firefox.