FEATURE: iframes on Firefox addon
This commit introduces a complete refactoring/cleanup of the Firefox
version, it supports Ember apps embedded into iframes on Firefox 
(Ember libraries detection and Devtool panel connection) and it has 
an optional tomster icon in the locationbar.

The tomster icon can be enabled/disabled from the addon preferences, 
and it will be visible when an Ember app is detected in the tab or 
in an iframe inside the tab).

Requirements:
- Firefox >= 30.0 (needed for the "evalInWindow" method used in content-script)

Tested on:
- http://jsfiddle.net/9zHj9/1/

to be merged into #182