Handle and display errors caused by CP calculation in the object inspector
Opening an object in the inspector means we eagerly calculate all of
that object's CPs. Sometimes CPs that don't expect to be calculated yet
may throw an error (for example CPs that depend on another property that
hasn't been set yet).

Previously, the object inspector just wouldn't
open, and the thrown errors were displayed in the app's console.

After this PR these errors are caught by the inspector and displayed
above the object's properties.