[Glimmer] The view tree now supports glimmer
There was a shift in concept from a view tree to a render node tree.
Previously, the template tree was made of a hierarchy of view instances.
After the Glimmer rewrite, view instances are optional and
secondary, and the main template representation is a render node tree.

This required a re-write of all of view-debug.js. Old code was left
untouched, so the current version supports both pre-glimmer and post-glimmer. We detect glimmer
by checking for the absence of `_childViews` on the top level view.

All code that touches private api (anything that has to do with render
nodes) needs to be moved to Ember core and tested.