Use the layout service to pass content height to lists
This allows us to also trigger events whenever the content height
changes. Events are needed to re-render the x-list content due to smoke
and mirrors not being aware that content height has changed and
requiring a rerender.

To reproduce the bug that this commit fixes:
- Open the inspector on an Ember.js app.
- Switch to the Console (or any other devtools tab).
- Reload the page (Inspector has now loaded but with height 0 since it's
  not visible).
- Visit the inspector.

Before this commit the view tree would only have one visible row due to smoke
and mirrors not being aware that content height has changed.