Remove all usage of jQuery from `/ember_debug` folder.
This folder is the one that actually runs along with the user's app, so in
theory this is enough to make the inspector work on apps without jQuery.
However on following PRs I'd like to completely remove jQuery from the
extension and ensure tests are run without it, so nobody inadvertently
reintroduces it by mistake.