Initial component tree implementation (#785)
* WIP First crack at component tree view

This first pass at a component tree is basically a stripped down copy of
the view tree.

Features:
- Clicking a component name opens the object inspector for that
  component.
- Hovering on a component will toggle the preview layer in the
  application DOM.
- Collapsible components

A ton of stuff is totally arbitrary right now but it's a starting point.

Filtering out non-component views isn't implemented yet. No tests yet
either as this is a rough spike.

* New down arrow SVG

* WIP First crack at styling component tree view

* WIP basic search

Filters items based on search result

TODO:
-  Reset indentation level to 0 for results but still allow for expansion
   ala react devtools

* Rename 'Component Tree' -> 'Components'

* Remove component selector from toolbar

* Select items in the DOM w/common selector button

This works just like the view tree, but slightly enhanced.  When a
component is 'pinned' it also shows up as selected in the tree.

KNOWN PROBLEMS:
- if the component is not technically visible it won't show up
- It'd probably be nice if the pinned component was scrolled into view
  in our tree view as well.

* Set view-debug options for individual routes

* Scroll to view icon for Component Tree

* Update padding to better clarify which elements are siblings in the tree

* Add basic scroll to element support

* Small code cleanup

* Don't try to scroll if element is not found

* Cache expanded state of tree nodes

When the view tree refreshes it's jarring if a node the user had
collapsed suddenly expands. We cache the expanded states and reconcile
the state of the new nodes against the cache to avoid it.

* Cache expanded state of tree nodes

When the view tree refreshes it's jarring if a node the user had
collapsed suddenly expands. We cache the expanded states and reconcile
the state of the new nodes against the cache to avoid it.

* Update component icon

* Fix iframe dropdown styles

* Consistent property name for search

* Add more tests

* eye svg to svg-jar

* Wrap svg-jar in div for click handling

For whatever reason trying to attach the event handler directly causes
an error, so we'll wrap it.

* Component tree clear search test

* Component tree item spacing and alignment fixes

* More component tree alignment tweaks

* Clean up tree flattening code and add comments

* Different styles and behavior for views in Component Tree

* Component tree expand and collapse all

Icons and placeholder actions

* Make all of component tree item clickable

* ExpandAll uses filteredArray rather than viewTree

The downstream properties only check that the viewTree's membership
has changed, not the individual properties.

* Expand parents for inspected components

* Update tree tests after styling change

* Respond to PR feedback

- Remove unused model hook
- Extract ComponentViewItem to a separate file