Use Ember.get to get _debugInfo (#847)
Resolves #791

I believe using Ember.get should fix this issue right @rwjblue?

It was this error:

```
Error message: Assertion Failed: You attempted to access the `_debugInfo` property (of <DS.PromiseObject:ember2256>).
Since Ember 3.1, this is usually fine as you no longer need to use `.get()`
to access computed properties. **However, in this case, the object in question
is a special kind of Ember object (a proxy). Therefore, it is still necessary
to use `.get('_debugInfo')` in this case.**
```