Use [class='ember-view'] selector to exclude anything with extra classes
This should exclude liquid-fire elements or various others that add classes to the application container ember-view.

If you were adding your own classes to the application container, this would break, but I think this supports the majority of cases, and application wrappers are disabled by a lot of people these days anyway.

Resolves #419