Remove unused code (#1061)
* Remove unused options

The inspector used to support filtering out components in the view
tab, and there are a lot of supporting code for that. Since the
view tab was removed, there is no longer any way to set this option,
and it is always true in practice.

* Remove unused `completeViewClass` property

This property contains the "full" class name string, something like
`<app@component:textarea::ember545>` as opposed to `textarea`. It is
only set but never read by anything.

* Remove unused `shouldShowView` method

This method is defined but never used. It is also the only thing
that reads the `isVirtual` property, which is an Ember 1.0-ism that
is no longer needed.

* Remove a lot (!) of unused code!

Most of the job of view-debug has already been moved to glimmer-tree,
leaving behind a lot of code that was never called.