Revert co-location for now
This seems to be causing build issues when running ember-try.

Co-location is not suppored in old versions of ember, so normally
addons that need to support old ember cannot use the co-location
feature. We are in a strange spot – the addon here is only used
internally for the "inspector app" only, basically just split out
for code organization purposes. Since the addon is only consumed
by one app (ours) and we know the app is on a recent ember version,
we should be able to use co-location as we please. However, we also
run ember-try on old ember versions. In these cases, we don't actually
care about the "app", we are only running ember-try to test the
ember-debug adpaters which do have to work on old ember versions and
does not have an UI (and so does not "use" co-location). However, we
still run the build on the old ember-source version, and this seems
to be causing problems.

EEXIST: file already exists, symlink '/tmp/broccoli-2882iUrxtLKjgtfM/out-281-funnel_funnel_ui_addon/ui/components/disclosure.js' -> '/tmp/broccoli-2882iUrxtLKjgtfM/out-282-broccoli_persistent_filter_template_compiler/ui/components/disclosure.js'

Co-authored-by: Robert Jackson <me@rwjblue.com>