Avoid errors when using older Ember versions. (#1102)
With ember-cli-htmlbars 4.2, there is a check to disable co-location
support on old ember versions (which, even when compiled, would fail
at runtime due to missing `setComponentTemplate`). This is causing
build failures when running ember-try with old Ember versions. Since
we don't actually run the app or its tests in ember-try, we can work
around the problem by skipping template compilation for now.

Co-authored-by: Godfrey Chan <godfreykfc@gmail.com>