Refactor acceptance tests
The main goal is to get the tests to more accurately reflect what
actually happen in real applications. Instead of mocking/reopening
the application "port", we now directly replace the adapter with a
`TestAdapter` that faithfully implements the required hooks the
same way that a "real" adapter does.

This allows us to write acceptance tests by only sending/responding
to messages in a way that is, as a far as the inspector app can is
concerned, indistinguishable from the "real" ember debug behaves in
the real world. The new API also allows us to be a bit more precise
in terms of the expectations (messages sent/received the correct
number of times, in the right order, etc).