Don't abuse timers (#1108)
Previously, the object inspector would schedule an infinite timer
that fires every 300ms to update the current object (it actually
fires even where there isn't a current object to update). This is
extremely inefficient, and it causes a lot of extra work in big
apps, since the timer starts a runloop which results in a full
top-down revalidation. The more appropriate thing to do is tap
into the runloop callbacks, since as far as Ember is concerned,
if there are no runloops, there are no observable changes.