Render the CHANGELOG for the current version
There are a few cases where the CHANGELOG on master may not match
the version you are actually running. For example, master can be
behind the stable branch's CHANGELOGs, or you could be loading and
running a nightly version. This patch try to fetch the CHANGELOG
from the correct Git tag to ensure things matches up correctly.