Bring back Node 8 support (for now) (#1116)
Revert some dependencies bump from #1096 because `ember-electron`
(probably the only npm consumer?) still support Node 8. This is
mostly a technicality, since they don't actually run the build and
only consume the artifacts in `dist` (included in the npm package).

We should probably find a better way for them to consume the app,
but in the mean time, this will allow them to consume the latest
inspector until Node 8 is officially EOL in a few weeks. We can
try this again when Ember CLI drops Node 8 support (likely 3.17).