Remove invalid and unecessary legacy {{#each}}
In Ember 3.14ish {{#each}} without an argument throws an exception. The
usage is not valid, however the tests also don't require this route to
exist. So remove it.

The upstream issue https://github.com/emberjs/ember.js/issues/18673
should still get resolved but there isn't a need to use the old usage
here.