Introduce `setupTestAdapter` hook (#1165)
While working on #1163, we encountered an issue with the test
adapter, where unmet expectations (when you set up a `respondTo`
for a test that was never triggered) did not cause the offending
test to fail. Instead, they became unhandled promise rejections
(i.e. "global errors") and fails the _next_ test that runs.

This is because we were using the `QUnit.{testStart,testDone}`
hooks as if they were a global version of `{before,after}Each`.
However, the global hooks are really intended for things like
test reporters and runs too late to affect things like the pass
or fail status of a test.

This commit introduces a `setupTestAdapter` function that uses
the normal (not global) APIs so that the failures are attached
to the correct test.