Upgrade to 3.18 (#1196)
There is a good chance this may fix issues like #1195. I haven't
dug in yet, but ember-source 3.17 has a bug with `{{#each}}` when
the same key is used more than once, which is the kind of thing
that could easily break vertical collection and is at least
visually consistent with the errors reported by users. It was
fixed in 3.18.1 via https://github.com/emberjs/ember.js/pull/18941
but the fix was never made it to 3.17 in time. So, either way, we
really should upgrade.