Sort container instances alphabetically
This commit uses <EmberTable>'s sorting capabilities to sort container
instances alphabetically be default. A sort indicator is shown in the
column header, and the header can be clicked to reverse the sort.

This PR adds test coverage for sorting and converts the template and
tests to use data-test- selectors instead of js- css classes.