Reset internal test helper state for EmberDebug tests.
Prior to this change we could not run any tests that leveraged the
`@ember/test-helpers` setup methods that ran _after_ the files in
`tests/ember_debug/**/*-test.js`. This was because the
`setupEIApp`/`destroyEIApp` helper functions would call `setApplication`
(to get a pristine state for their own tests), but would not clean up
after themselves (calling `setApplication` with the original value).

This change is a minimal patch to fix the fundamental issue (and adding
a single integration test to the mix to ensure it doesn't regress), but
a larger refactor is in order to simplify these setupEI/destroyEI test
helpers.