Remove private API usage to set moduleName in view debug tests.
This leverages the `CallExpression` form of the inline `hbs` compiler so
that we can provide a custom module name and avoid having to muck with
the private internal state of the compiled template itself (e.g.
`__meta`).

This mechanism is exactly the same as what is used by
`ember-cli-htmlbars` when compiling standalone templates, and is much
more resilient to internal private API changes in Ember itself.