Ensure `valueForTag` is passed an actual `Tag` (#1299)
`@glimmer/validator` only accepts a `Tag` instance as an argument
but we are passing the object + property name (basically we are
treating it like `tagForProperty`). Recent versions of ember-source
(that have updated versions of glimmer-vm) actually use the passed
in arguments to return the revision so we are hitting an error
(since the `object` we pass is not a `Tag`).