Update publish-chrome action to use node 14.X
The latest version of chrome-webstore-upload-cli uses optional chaining (?. syntax), which is not supported until Node 14 (without polyfills). Updated the publish-chrome action to use node 14. Left the other actions on 12 for now because I'm not too comfortable with github actions or this repo.

Note: I was unable to test this locally - I am not sure how to run/test GitHub actions on local, but I fairly certain this should work.