Fix chrome publishing
It seems like Google may have changed some of their APIs and also
the the `chrome-webstore-upload-cli` had some changes and does not
require the client secret anymore. I followed their guide to remake
the tokens. With the new process it seems like we need two sets of
tokens for the two extensions. If this all works correctly then I
will delete the old tokens from GitHub once this is merged.