deprecations: fixup instances of no-inline-styles
ember-template-lint's no-inline-styles rule forbids using inline styles
as they are often harder to maintain and can make the overall size of
the project bigger. It recommends using css classes instead.

Let's fixup any instances of this issue in the codebase and then
re-enable the rule in our template-lintrc so that the linter will help
us avoid adding more instances of this in the future.