deprecations: resolve instances of this-property-fallback (#1973)
It is currently possible to reference properties on a component without
a preceding `this`. This style of looking up properties is known as
"property fallback", and has the potential to collide with other types
of values. The alternative way to lookup properties is with a preceding
`this`. This style does not have any ambiguity, since it's clear that
we're looking up the property on this instance of the component, and not
the global helper/component. Property fallback has been deprecated in
favor of this style in general.

As such, let's work through all of our templates in order to eliminate any
instances of this issue. Finally, let's stop silencing violations of this
deprecation in our deprecation-workflow config file. This will ensure that
future violations are a little easier to catch.