template-lintrc: enable 'table-groups' rule
Right now, we disable the 'table-groups' linting rule. This rule helps
to ensure that table rows are grouped into one of: thead, tbody, tfoot,
helps avoid relying upon a (possibly to be deprecated) feature of
glimmer that auto-inserts these tags, and enforces that children of
<table> are ordered correctly. As such, it would be ideal to have this
enabled.

We don't have any outstanding issues that would cause errors to be
thrown by the linter by enabling this, so let's go ahead and do so.