eslintrc: enable 'no-actions-hash' lint rule (#1988)
As of Ember Octane, the actions hash and {{action}} modifier and helper
are no longer needed.  We should instead use the @action decorator or
`foo: action(function() {}))` syntax instead.

Right now, we've disabled the 'no-actions-hash' eslint rule. We do not
currently have any issues of this nature, so let's go ahead and
re-enable the rule so that the linter will warn us against creating new
issues of this type going forward.