render-item: fixup no-curly-component-invocation (#1987)
There are two ways to invoke a component in a template: curly component
syntax ({{my-component}}), and angle bracket syntax (<MyComponent />).
The difference between them is syntactical. We should favour angle
bracket syntax as it improves readability of templates, i.e.
disambiguates components from helpers, and is also the future direction
Ember is going with the Octane Edition.

As such, let's fix up the render-item template so that it follows this
syntax and then remove the manual disabling of this linting rule so that
the linter will warn us against adding more issues of this type in the
future.