list-cell: convert to glimmer (#1998)
Right now, the list-cell component is a classic component. For
this particular component, migrating to Glimmer will allow us to take
advantage of the simpler API, no wrapper element, and namespaced
arguments.

As such, let's go ahead and migrate over list-cell so that it
extends glimmer/component instead of ember/component.