Deprecation toolbar clear btn does not clear deprecations (#2008)
* deprecations: add a test case for toolbar clear btn

Let's add a test case which verifies that the clear button on the
toolbar will clear all of the deprecations in the deprecation pane when
clicked. This should help us avoid future regressions to this feature.

* deprecations: fixup deprecations accessor

When we try to clear the currentModel, an error is thrown as this method
does not exist. Let's instead just clear the deprecations directly on
the controller, similarly to how we add deprecations to the controller
in `deprecationsAdded()`.`

* side-nav: re-render items when args change

Right now, we won't re-render the side-nav items after the constructor.
Some of the variables used internally come from args that can change at
runtime, so it'd be nice if the side-nav would re-render when those
updates do occur.