update publish remove token for checkout
the token was there to support a flow with no/mininimal changes to release-plan.
release plan would do its npm release. create tag & git release.
the git tag would then trigger the workflow to do the chrome/firefox release.
to support this flow a PAT token was required. as otherwise a git tag would not trigger a new workflow